requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
datasets==2.20.0
Deprecated==1.2.14
dill==0.3.8
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.52.4
fredapi==0.5.2
frozenlist==1.4.1
fsspec==2024.5.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.23.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
huggingface-hub==0.24.3
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
multidict==6.0.5
multiprocess==0.70.16
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
rsa==4.9
seaborn==0.13.2
six==1.16.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
xlrd==2.0.1
xxhash==3.4.1
yarl==1.9.4
